+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/scg/PolarisReactiveLoadBalancerClientFilterTest.java
@@ -59,7 +59,7 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 /**
- * Test for ${@link PolarisReactiveLoadBalancerClientFilter}
+ * Test for ${@link PolarisReactiveLoadBalancerClientFilter}.
  *@author lepdou 2022-07-04
  */
 @RunWith(MockitoJUnitRunner.class)
@@ -97,7 +97,7 @@ public class PolarisReactiveLoadBalancerClientFilterTest {
 		Map<String, String> transitiveLabels = new HashMap<>();
 		transitiveLabels.put("t1", "v1");
 		transitiveLabels.put("t2", "v2");
-		when(metadataContext.getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE)).thenReturn(transitiveLabels);
+		when(metadataContext.getTransitiveMetadata()).thenReturn(transitiveLabels);
 
 		mockedMetadataContextHolder = Mockito.mockStatic(MetadataContextHolder.class);
 		mockedMetadataContextHolder.when(MetadataContextHolder::get).thenReturn(metadataContext);