+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [feat:report the labels when using RestTemplate.](https://github.com/Tencent/spring-cloud-tencent/pull/629)
 - [Bugfix: fix throw npe when router context is null](https://github.com/Tencent/spring-cloud-tencent/pull/635)
 - [fix:cancel reporting useless metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/639)
+- [Optimize:optimize transfer](https://github.com/Tencent/spring-cloud-tencent/pull/636)